some refactoring and math corrections (vec2 and rect)

+++ include/Alien.hpp
@@ -3,6 +3,7 @@
 
 #define OBJECT_SPEED 100
 
+#include "Sprite.hpp"
 #include "Component.hpp"
 #include "Vec2.hpp"
 #include <queue>
@@ -15,13 +16,13 @@ class Alien : public Component {
                 enum ActionType {MOVE, SHOOT};
                 Vec2 pos;
                 ActionType type;
-                Action(ActionType type, float x, float y): type(type), pos({(int)x,(int)y}) {}
+                Action(ActionType type, float x, float y): type(type), pos({(float)x,(float)y}) {}
         };
         Vec2 speed;
         int hp;
         std::queue<Action> taskQueue;
         std::vector<std::weak_ptr<GameObject>> minionArray;
-    
+            
     public:
         Alien(GameObject& associated, int nMinions);
         ~Alien();
